Refactor: 불필요한 타입 파일 변경
