+++ src/features/profiles/model/profile.type.ts
@@ -1,0 +1,1 @@
+export type BasicQAAnswers = Record<string, any>;
